+++ ScreenedNpole2d.h
@@ -99,7 +99,7 @@ class ScreenedNpole2d
 
 	ScreenedNpole2d(double xPos, double yPos, double radius, std::vector<double>& str, double laplaceCoeff, double screenCoeff)
 	{
-	this->momentCount = str.size();
+	this->momentCount = (long)str.size();
 	if(this->momentCount > 6) this->momentCount = 6;
 
 	int maxOrder;
